Exclude inactive people on PeopleList
refs: MBL-13753
affects: Student, Teacher
release note: Fixed inactive users showing up in people list
